Refactor: Upgrade goal details with interactive charts and statistical insights
This commit enhances the `GoalDetailScreen` by replacing the static summary card with a paginated interface featuring historical charts and statistical analysis. It also migrates internal chart data structures to use `Double` for improved financial precision.

- **Goal Detail Screen Enhancements:**
    - **Interactive Paging**: Integrated a `HorizontalPager` within the `GoalSummaryCard` to allow users to swipe between basic goal info, historical performance charts, and statistical summaries.
    - **Visual Analytics**: Implemented `GoalLineChartSummary` using `VicoLineChart` to visualize the trend of achievement amounts over time.
    - **Data Science Insights**: Introduced a `GoalStatistic` component that calculates and displays key metrics, including success rates, average achievement values, variance, and standard deviation.
    - **Component Refactoring**: Extracted core goal details (target progress, frequency, and deadlines) into a dedicated `GoalInfo` component for better maintainability.

- **Chart Data Precision:**
    - **Type Migration**: Updated `ChartData` and `GetYesterdayChartDataUseCase` to use `DoubleArray` and `Double` lists instead of `Int`. This ensures that fractional financial amounts are accurately represented in charts without truncation.

- **UI and Navigation:**
    - Added navigation controls (ChevronLeft/Right) and state-aware icons (Info, ShowChart, Summarize) to the goal summary header to improve pager discoverability.
    - Refined the progress bar styling and updated labels to clearly indicate the percentage of the target reached.

+++ app/src/main/java/com/example/moneytracker/ui/components/charts/collections/ChartData.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 
 data class ChartData(
     val x: List<Double> = listOf(),
-    val y: List<Int> = listOf(),
+    val y: List<Double> = listOf(),
     val color: Color = Color.Gray,
     // Optional label shown in legends. Keep default null to avoid breaking existing callers.
     val label: String? = null
